added cors and updated booking authenciation

+++ backend/controller/booking.js
@@ -1,18 +1,19 @@
 const asyncHandler=require('../middlewares/async')
 const Booking=require('../models/Booking');
 const ErrorResponce = require('../utils/ErrorResponce');
+//const User=require('../models/User')
 
 //CREATE BOOKING
 //private @customers and admin only
 exports.bookNow=asyncHandler(async(req,res,next)=>{
 
-const tour= await Booking.create(req.body);
+ req.body.userId=req.user.id;   
+const book= await Booking.create(req.body);
 
-if (!tour){
+if (!book){
     return next(new ErrorResponce(`Error in creating the tour`,406));
 }
-
-res.status(201).json({success: true,data:tour});
+res.status(201).json({success: true,data:book});
 
 });
 
@@ -21,7 +22,12 @@ res.status(201).json({success: true,data:tour});
 
 exports.getAllBookings=asyncHandler(async(req,res,next)=>{
 
-res.status(200).json(res.advancedResults);
+    const tour= await Booking.findAll();
+
+    if(!tour){
+        return next(new ErrorResponce(`Error in retrieving the tour`,406));
+    }
+    res.status(200).json({success: true,data:tour});
 
 });
 
@@ -41,14 +47,22 @@ exports.getSingleBooking=asyncHandler(async(req,res,next)=>{
 //PUT Update Booking
 //Private @user
 exports.updateBooking=asyncHandler(async(req,res,next)=>{
+    let tour=await Booking.findByPk(req.params.id);
 
-    const tour=await Booking.update(req.body,{
+
+    if(tour.id.toString()!==req.user.id || req.user.role!=='admin'){
+        return next(new ErrorResponce(`User with ${req.user.id} not authorised to access`,406));
+    }
+
+      tour=await Booking.update(req.body,{
         where:{id:req.params.id}
     });
 
     if(!tour){
         return next(new ErrorResponce(`Error in updating the item`,406));
     }
+
+    
     res.status(200).json({success: true,data:tour});
 
 });
@@ -58,7 +72,15 @@ exports.updateBooking=asyncHandler(async(req,res,next)=>{
 
 exports.deleteSingleBooking=asyncHandler(async(req,res,next)=>{
 
-    const tour=await Booking.destroy({where:{id:req.params.id}})
+    let tour=await Booking.findByPk(req.params.id);
+
+
+    if(tour.id.toString()!==req.user.id || req.user.role!=='admin'){
+        return next(new ErrorResponce(`User with ${req.user.id} not authorised to access`,406));
+    }
+
+
+     tour=await Booking.destroy({where:{id:req.params.id}})
 
     if(!tour){
         return next(new ErrorResponce(`Error in deleting the item`,406));
@@ -70,8 +92,13 @@ exports.deleteSingleBooking=asyncHandler(async(req,res,next)=>{
 //DELETE all Booking
 //Private @user
 exports.deleteAllBooking=asyncHandler(async(req,res,next)=>{
+    let tour=await Booking.findByPk(req.params.id);
 
-    const tour=await Booking.destroy({where:{},truncate:true});
+
+    if(tour.id.toString()!==req.user.id || req.user.role!=='admin'){
+        return next(new ErrorResponce(`User with ${req.user.id} not authorised to access`,406));
+    }
+     tour=await Booking.destroy({where:{},truncate:true});
 
     if(!tour){
         return next(new ErrorResponce(`Error in deleting the item`,406));
